testing

+++ Sort-Bot/CaseProcessing.h
@@ -5,7 +5,7 @@ using namespace cv;
 class CaseProcessor {
 	static void OpenCaseAndTakeScreen(std::shared_ptr<PointCaseInStash> ptr_PCIS);
 
-	static void MatchingCaseInCase(Mat& MatScreen, int page, POINT parentCasePoints);
+	static void MatchingCaseInCase(Mat& MatScreen, uint8_t page, POINT parentCasePoints);
 
 	static void cleanUpVectorCase();
